ophidian/requirements/ophidian.txt
appdirs==1.4.3
attrs==16.3.0
click==6.7
packaging==16.8
pyparsing==2.2.0
six==1.10.0
